update-table-id-and-fix-service

+++ PaymentModule/src/main/java/com/cs490/shoppingCart/PaymentModule/repository/TransactionRepository.java
@@ -7,9 +7,9 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface TransactionRepository extends JpaRepository<Transaction, Integer> {
-    List<Transaction> findTransactionsByUserId(Integer id);
-    Transaction findTransactionByOrderId(Integer id);
+public interface TransactionRepository extends JpaRepository<Transaction, Long> {
+    List<Transaction> findTransactionsByUserId(Long id);
+    Transaction findTransactionByOrderId(Long id);
     Transaction findFirstByCardNumberOrderByIdDesc(String cardNumber);
 
 }
